Renamings Explanation => ElExplanation + rework attributes
