--- ui-components/src/stories/explanation/Explanation.stories.ts
@@ -1,42 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/vue3'
-import Explanation from '@/components/explanation/Explanation.vue'
-import type {MultipleChoiceResponse, OpenEndedResponse} from "@/models/Response";
-
-
-const meta = {
-  title: 'explanation/Explanation',
-  component: Explanation,
-  tags: ['autodocs'],
-} satisfies Meta<typeof Explanation>
-
-export default meta
-
-type Story = StoryObj<typeof meta>
-
-export const Default: Story = {
-  args: {
-    answer: {
-      id: 1,
-      questionType: 'MultipleChoice',
-      explanation: 'Mon explication',
-      choices: [1,3],
-      trust: "",
-    } satisfies MultipleChoiceResponse,
-    grade: 2,
-    numberOfPeerReview: 1,
-    providedByTeacher: false,
-  },
-}
-export const OpenQuestion: Story = {
-  args: {
-    answer: {
-      id: 2,
-      questionType: 'OpenEnded',
-      explanation: 'Ma Réponse',
-      trust: "",
-    } satisfies OpenEndedResponse,
-    grade: 2,
-    numberOfPeerReview: 1,
-    providedByTeacher: false,
-  },
-}
